changed CMakeLists.txt to create separate executable for each component test
